feat(DB): Add new v_abgegrenzte_interessenbindung view
Adds a new database view `v_abgegrenzte_interessenbindung` which in contrast to the `interessebindung` table has the columns `von_effektiv` and `bis_effektiv` to represent the start and end date for the validity period (as opposed to `von` and `bis`). These dates are never `null` and when the end date is yet unknown (open-ended) the `bis_effektiv` will be represented by the date 9999-12-31.

The value for `von_effektiv` will correspond to `von` when given and otherwise to `freigabe_datum` or `parlamentarier.im_rat_seit`, in case the `interessebindung` row was created as part of the initial research of the politician.

Also note that in case a politician was elected for two non-consecutive periods of time (cf. `parlamentarier.ratsunterbruch_von` and `parlamentarier.ratsunterbruch_bis`), then the same `interessebindung` object may (depending on `von` and `bis`) end up being represented by multiple rows in this view (both with the same value for `interessebindung_id`).